+++ arch/armv7m/kernel/inc/kcoredefs.h
@@ -4,7 +4,7 @@
 /** RK0 - The Embedded Real-Time Kernel '0'                                   */
 /** (C) 2026 Antonio Giacomelli <dev@kernel0.org>                             */
 /**                                                                           */
-/** VERSION: 0.9.18                                                           */
+/** VERSION: 0.9.19                                                           */
 /**                                                                           */
 /** You may obtain a copy of the License at :                                 */
 /** http://www.apache.org/licenses/LICENSE-2.0                                */
@@ -83,6 +83,13 @@ static inline void kExitCR(unsigned state)
 
 RK_FORCE_INLINE static inline unsigned kIsISR(void)
 {
+#if defined(RK_QEMU_UNIT_TEST)
+    extern volatile unsigned RK_gQemuTestForceIsr;
+    if (RK_gQemuTestForceIsr != 0U)
+    {
+        return (1U);
+    }
+#endif
     unsigned ipsr_value;
     RK_ASM("MRS %0, IPSR" : "=r"(ipsr_value));
     return (ipsr_value);