+++ core/inc/kversion.h
@@ -4,7 +4,7 @@
 /** RK0 - The Embedded Real-Time Kernel '0'                                   */
 /** (C) 2026 Antonio Giacomelli <dev@kernel0.org>                             */
 /**                                                                           */
-/** VERSION: 0.9.18                                                           */
+/** VERSION: 0.9.19                                                           */
 /**                                                                           */
 /** You may obtain a copy of the License at :                                 */
 /** http://www.apache.org/licenses/LICENSE-2.0                                */
@@ -29,7 +29,7 @@ struct RK_gKversion
 
 #define RK_VERSION_MAJOR 0
 #define RK_VERSION_MINOR 9
-#define RK_VERSION_PATCH 18
+#define RK_VERSION_PATCH 19
 
 
 extern struct RK_gKversion const RK_gKversion;